[Updated] Packet for crafting and resource containers [Fixed] Assorted Crafting implementation bugs SQL UPDATE REQUIRED
git-svn-id: http://svn2.assembla.com/svn/swgemu/core3/trunk/MMOCoreORB@1483 c3d1530f-68f5-4bd0-87dc-8ef779617e40

+++ src/server/zone/managers/combat/CombatManager.cpp
@@ -377,12 +377,12 @@ void CombatManager::handleAreaAction(CreatureObject* creature, TangibleObject* t
 		if (skill->isThrowSkill()) {
 			ThrowableWeapon* throwWeapon = ((ThrowAttackTargetSkill*)skill)->getThrowableWeapon(creature,actionModifier);
 			if (throwWeapon != NULL && throwWeapon->isGrenade())
-				areaRange = throwWeapon->getArea();
+				areaRange = (int)throwWeapon->getArea();
 		}
 
 		if (skill->isHeavyWeaponSkill()) {
 			HeavyRangedWeapon* heavyWeeapon = ((FireHeavyWeaponAttackTarget*)skill)->getHeavyRangedWeapon(creature,actionModifier);
-			areaRange = heavyWeeapon->getArea();
+			areaRange = (int)heavyWeeapon->getArea();
 		}
 
 		bool rangeArea = skill->isThrowSkill() || skill->isHeavyWeaponSkill();
